+++ rust/client/src/ui/theme/mod.rs
@@ -22,6 +22,7 @@ pub mod row;
 pub mod rule;
 pub mod scrollable;
 pub mod space;
+pub mod svg;
 pub mod text;
 pub mod text_input;
 pub mod tooltip;