+++ rust/client/src/ui/widget/data.rs
@@ -37,7 +37,7 @@ pub struct ComponentWidgets<'b> {
     pub root_widget: &'b Option<Arc<RootWidget>>,
     pub state: &'b HashMap<UiWidgetId, ComponentWidgetState>,
     pub plugin_id: PluginId,
-    pub images: &'b HashMap<UiWidgetId, Vec<u8>>,
+    pub data: &'b HashMap<UiWidgetId, Vec<u8>>,
 }
 
 impl<'b> ComponentWidgets<'b> {
@@ -45,13 +45,13 @@ impl<'b> ComponentWidgets<'b> {
         root_widget: &'b Option<Arc<RootWidget>>,
         state: &'b HashMap<UiWidgetId, ComponentWidgetState>,
         plugin_id: PluginId,
-        images: &'b HashMap<UiWidgetId, Vec<u8>>,
+        data: &'b HashMap<UiWidgetId, Vec<u8>>,
     ) -> ComponentWidgets<'b> {
         Self {
             root_widget,
             state,
             plugin_id,
-            images,
+            data,
         }
     }