+++ rust/client/src/ui/widget_container.rs
@@ -26,7 +26,7 @@ use crate::ui::AppMsg;
 pub struct PluginWidgetContainer {
     root_widget: Option<Arc<RootWidget>>,
     state: HashMap<UiWidgetId, ComponentWidgetState>,
-    images: HashMap<UiWidgetId, Vec<u8>>,
+    data: HashMap<UiWidgetId, Vec<u8>>,
     plugin_id: Option<PluginId>,
     plugin_name: Option<String>,
     entrypoint_id: Option<EntrypointId>,
@@ -38,7 +38,7 @@ impl PluginWidgetContainer {
         Self {
             root_widget: None,
             state: HashMap::new(),
-            images: HashMap::new(),
+            data: HashMap::new(),
             plugin_id: None,
             plugin_name: None,
             entrypoint_id: None,
@@ -61,7 +61,7 @@ impl PluginWidgetContainer {
     pub fn replace_view(
         &mut self,
         container: Arc<RootWidget>,
-        images: HashMap<UiWidgetId, Vec<u8>>,
+        data: HashMap<UiWidgetId, Vec<u8>>,
         plugin_id: &PluginId,
         plugin_name: &str,
         entrypoint_id: &EntrypointId,
@@ -73,7 +73,7 @@ impl PluginWidgetContainer {
         self.plugin_name = Some(plugin_name.to_string());
         self.entrypoint_id = Some(entrypoint_id.clone());
         self.entrypoint_name = Some(entrypoint_name.to_string());
-        self.images = images;
+        self.data = data;
 
         // use new state with values from old state but only widget ids which exists in new state
         // so this way we use already existing values but remove state for removed widgets
@@ -96,7 +96,7 @@ impl PluginWidgetContainer {
         self.root_widget = Some(container);
 
         if first_open {
-            ComponentWidgets::new(&mut self.root_widget, &mut self.state, plugin_id.clone(), &self.images).first_open()
+            ComponentWidgets::new(&mut self.root_widget, &mut self.state, plugin_id.clone(), &self.data).first_open()
         } else {
             AppMsg::Noop
         }
@@ -113,7 +113,7 @@ impl PluginWidgetContainer {
         plugin_view_state: &PluginViewState,
         action_shortcuts: &HashMap<String, PhysicalShortcut>,
     ) -> Element<'a, ComponentWidgetEvent> {
-        ComponentWidgets::new(&self.root_widget, &self.state, self.get_plugin_id(), &self.images).render_root_widget(
+        ComponentWidgets::new(&self.root_widget, &self.state, self.get_plugin_id(), &self.data).render_root_widget(
             plugin_view_state,
             self.entrypoint_name.as_ref(),
             action_shortcuts,
@@ -121,60 +121,60 @@ impl PluginWidgetContainer {
     }
 
     pub fn render_inline_root_widget<'a>(&self) -> Element<'a, ComponentWidgetEvent> {
-        ComponentWidgets::new(&self.root_widget, &self.state, self.get_plugin_id(), &self.images)
+        ComponentWidgets::new(&self.root_widget, &self.state, self.get_plugin_id(), &self.data)
             .render_root_inline_widget(self.plugin_name.as_ref(), self.entrypoint_name.as_ref())
     }
 
     pub fn append_text(&mut self, text: &str) -> Task<AppMsg> {
         let plugin_id = self.get_plugin_id();
-        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.images).append_text(text)
+        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.data).append_text(text)
     }
 
     pub fn backspace_text(&mut self) -> Task<AppMsg> {
         let plugin_id = self.get_plugin_id();
-        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.images).backspace_text()
+        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.data).backspace_text()
     }
 
     pub fn focus_search_bar(&self, widget_id: UiWidgetId) -> Task<AppMsg> {
         let plugin_id = self.get_plugin_id();
-        ComponentWidgets::new(&self.root_widget, &self.state, plugin_id, &self.images).focus_search_bar(widget_id)
+        ComponentWidgets::new(&self.root_widget, &self.state, plugin_id, &self.data).focus_search_bar(widget_id)
     }
 
     pub fn toggle_action_panel(&mut self) {
         let plugin_id = self.get_plugin_id();
-        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.images).toggle_action_panel()
+        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.data).toggle_action_panel()
     }
 
     pub fn get_action_ids(&self) -> Vec<UiWidgetId> {
-        ComponentWidgets::new(&self.root_widget, &self.state, self.get_plugin_id(), &self.images).get_action_ids()
+        ComponentWidgets::new(&self.root_widget, &self.state, self.get_plugin_id(), &self.data).get_action_ids()
     }
 
     pub fn get_focused_item_id(&self) -> Option<String> {
-        ComponentWidgets::new(&self.root_widget, &self.state, self.get_plugin_id(), &self.images).get_focused_item_id()
+        ComponentWidgets::new(&self.root_widget, &self.state, self.get_plugin_id(), &self.data).get_focused_item_id()
     }
 
     pub fn get_action_panel(&self, action_shortcuts: &HashMap<String, PhysicalShortcut>) -> Option<ActionPanel> {
-        ComponentWidgets::new(&self.root_widget, &self.state, self.get_plugin_id(), &self.images)
+        ComponentWidgets::new(&self.root_widget, &self.state, self.get_plugin_id(), &self.data)
             .get_action_panel(action_shortcuts)
     }
 
     pub fn focus_up(&mut self) -> Task<AppMsg> {
         let plugin_id = self.get_plugin_id();
-        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.images).focus_up()
+        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.data).focus_up()
     }
 
     pub fn focus_down(&mut self) -> Task<AppMsg> {
         let plugin_id = self.get_plugin_id();
-        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.images).focus_down()
+        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.data).focus_down()
     }
 
     pub fn focus_left(&mut self) -> Task<AppMsg> {
         let plugin_id = self.get_plugin_id();
-        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.images).focus_left()
+        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.data).focus_left()
     }
 
     pub fn focus_right(&mut self) -> Task<AppMsg> {
         let plugin_id = self.get_plugin_id();
-        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.images).focus_right()
+        ComponentWidgetsMut::new(&mut self.root_widget, &mut self.state, plugin_id, &self.data).focus_right()
     }
 }